Refactor submodule update logic for git operations
Consolidates submodule update logic into a generic updateSubmodulesAfterOperation function in submodule.ts, replacing operation-specific implementations in checkout.ts and pull.ts. Also moves and generalizes clampProgress to progress.ts for reuse. This improves code maintainability and reduces duplication.
