+++ app/src/lib/git/checkout.ts
@@ -1,11 +1,10 @@
 import { git, IGitStringExecutionOptions } from './core'
 import { Repository } from '../../models/repository'
 import { Branch, BranchType } from '../../models/branch'
-import { ICheckoutProgress } from '../../models/progress'
+import { clampProgress, ICheckoutProgress } from '../../models/progress'
 import {
   CheckoutProgressParser,
   executionOptionsWithProgress,
-  IGitOutput,
 } from '../progress'
 import { AuthenticationErrors } from './authentication'
 import {
@@ -16,22 +15,11 @@ import { WorkingDirectoryFileChange } from '../../models/status'
 import { ManualConflictResolution } from '../../models/manual-conflict-resolution'
 import { CommitOneLine, shortenSHA } from '../../models/commit'
 import { IRemote } from '../../models/remote'
+import { updateSubmodulesAfterOperation } from './submodule'
 
 export type ProgressCallback = (progress: ICheckoutProgress) => void
 
 const CheckoutStepWeight = 0.9
-
-function clampProgress(
-  minimum: number,
-  maximum: number,
-  progressCallback: ProgressCallback
-): ProgressCallback {
-  return (progress: ICheckoutProgress) =>
-    progressCallback({
-      ...progress,
-      value: minimum + progress.value * (maximum - minimum),
-    })
-}
 
 function getCheckoutArgs(progressCallback?: ProgressCallback) {
   return ['checkout', ...(progressCallback ? ['--progress'] : [])]
@@ -98,109 +86,6 @@ async function getCheckoutOpts(
 }
 
 /**
- * Update submodules after a checkout operation.
- *
- * @param repository - The repository in which to update submodules
- * @param title - The title to use for progress reporting
- * @param target - The target branch/commit being checked out
- * @param currentRemote - The current remote for environment setup
- * @param progressCallback - An optional function which will be invoked
- *                           with information about the current progress
- *                           of the submodule update operation.
- */
-async function updateSubmodulesAfterCheckout(
-  repository: Repository,
-  title: string,
-  target: string,
-  currentRemote: IRemote | null,
-  progressCallback: ProgressCallback | undefined,
-  allowFileProtocol: boolean
-): Promise<void> {
-  const opts: IGitStringExecutionOptions = {
-    env: await envForRemoteOperation(
-      getFallbackUrlForProxyResolve(repository, currentRemote)
-    ),
-    expectedErrors: AuthenticationErrors,
-  }
-
-  const args = [
-    ...(allowFileProtocol ? ['-c', 'protocol.file.allow=always'] : []),
-    'submodule',
-    'update',
-    '--init',
-    '--recursive',
-  ]
-
-  if (!progressCallback) {
-    await git(args, repository.path, 'updateSubmodules', opts)
-    return
-  }
-
-  // Initial progress
-  progressCallback({
-    kind: 'checkout',
-    title,
-    description: 'Updating submodules',
-    value: 0,
-    target,
-  })
-
-  const kind = 'checkout'
-
-  let submoduleEventCount = 0
-
-  const progressOpts = await executionOptionsWithProgress(
-    { ...opts, trackLFSProgress: true },
-    {
-      parse(line: string): IGitOutput {
-        if (
-          line.match(/^Submodule path (.)+?: checked out /) ||
-          line.startsWith('Cloning into ')
-        ) {
-          submoduleEventCount += 1
-        }
-
-        return {
-          kind: 'context',
-          text: `Updating submodules: ${line}`,
-          // Math taken from https://math.stackexchange.com/a/2323106
-          // We do this to fake a progress that slows down as we process more
-          // events, as we don't know how many submodules there are upfront, or
-          // what does git have to do with them (cloning, just checking them
-          // out...)
-          percent: 1 - Math.exp(-submoduleEventCount * 0.25),
-        }
-      },
-    },
-    progress => {
-      const description =
-        progress.kind === 'progress' ? progress.details.text : progress.text
-
-      const value = progress.percent
-
-      progressCallback({
-        kind,
-        title,
-        description,
-        value,
-        target,
-      })
-    }
-  )
-
-  await git(args, repository.path, 'updateSubmodules', progressOpts)
-
-  // Final progress
-  progressCallback({
-    kind,
-    title,
-    description: 'Submodules updated',
-    value: 1,
-    target,
-  })
-}
-
-/**
  * Check out the given branch.
  *
  * @param repository - The repository in which the branch checkout should
@@ -239,14 +124,19 @@ export async function checkoutBranch(
   await git(args, repository.path, 'checkoutBranch', opts)
 
   // Update submodules after checkout
-  await updateSubmodulesAfterCheckout(
+  await updateSubmodulesAfterOperation(
     repository,
-    title,
-    branch.name,
     currentRemote,
     progressCallback
-      ? clampProgress(CheckoutStepWeight, 1, progressCallback)
+      ? clampProgress<ICheckoutProgress>(
+          CheckoutStepWeight,
+          1,
+          progressCallback
+        )
       : undefined,
+    'checkout',
+    title,
+    branch.name,
     allowFileProtocol
   )
 
@@ -295,14 +185,19 @@ export async function checkoutCommit(
   await git(args, repository.path, 'checkoutCommit', opts)
 
   // Update submodules after checkout
-  await updateSubmodulesAfterCheckout(
+  await updateSubmodulesAfterOperation(
     repository,
-    title,
-    target,
     currentRemote,
     progressCallback
-      ? clampProgress(CheckoutStepWeight, 1, progressCallback)
+      ? clampProgress<ICheckoutProgress>(
+          CheckoutStepWeight,
+          1,
+          progressCallback
+        )
       : undefined,
+    'checkout',
+    title,
+    target,
     allowFileProtocol
   )
