+++ app/src/lib/git/pull.ts
@@ -1,31 +1,16 @@
 import { git, gitRebaseArguments, IGitStringExecutionOptions } from './core'
 import { Repository } from '../../models/repository'
-import { IPullProgress } from '../../models/progress'
-import {
-  PullProgressParser,
-  executionOptionsWithProgress,
-  IGitOutput,
-} from '../progress'
+import { clampProgress, IPullProgress } from '../../models/progress'
+import { PullProgressParser, executionOptionsWithProgress } from '../progress'
 import { IRemote } from '../../models/remote'
 import {
   envForRemoteOperation,
   getFallbackUrlForProxyResolve,
 } from './environment'
 import { getConfigValue } from './config'
+import { updateSubmodulesAfterOperation } from './submodule'
 
 const PullStepWeight = 0.9
-
-function clampProgress(
-  minimum: number,
-  maximum: number,
-  progressCallback: (progress: IPullProgress) => void
-): (progress: IPullProgress) => void {
-  return (progress: IPullProgress) =>
-    progressCallback({
-      ...progress,
-      value: minimum + progress.value * (maximum - minimum),
-    })
-}
 
 async function getPullArgs(
   repository: Repository,
@@ -39,105 +24,6 @@ async function getPullArgs(
     ...(progressCallback ? ['--progress'] : []),
     remote,
   ]
-}
-
-/**
- * Update submodules after a pull operation.
- *
- * @param repository - The repository in which to update submodules
- * @param remote - The remote that was pulled from
- * @param progressCallback - An optional function which will be invoked
- *                           with information about the current progress
- *                           of the submodule update operation.
- */
-async function updateSubmodulesAfterPull(
-  repository: Repository,
-  remote: IRemote,
-  progressCallback: ((progress: IPullProgress) => void) | undefined,
-  allowFileProtocol: boolean
-): Promise<void> {
-  const opts: IGitStringExecutionOptions = {
-    env: await envForRemoteOperation(
-      getFallbackUrlForProxyResolve(repository, remote)
-    ),
-  }
-
-  const args = [
-    ...(allowFileProtocol ? ['-c', 'protocol.file.allow=always'] : []),
-    'submodule',
-    'update',
-    '--init',
-    '--recursive',
-  ]
-
-  if (!progressCallback) {
-    await git(args, repository.path, 'updateSubmodules', opts)
-    return
-  }
-
-  // Initial progress
-  const title = `Pulling ${remote.name}`
-  progressCallback({
-    kind: 'pull',
-    title,
-    description: 'Updating submodules',
-    value: 0,
-    remote: remote.name,
-  })
-
-  const kind = 'pull'
-
-  let submoduleEventCount = 0
-
-  const progressOpts = await executionOptionsWithProgress(
-    { ...opts, trackLFSProgress: true },
-    {
-      parse(line: string): IGitOutput {
-        if (
-          line.match(/^Submodule path (.)+?: checked out /) ||
-          line.startsWith('Cloning into ')
-        ) {
-          submoduleEventCount += 1
-        }
-
-        return {
-          kind: 'context',
-          text: `Updating submodules: ${line}`,
-          // Math taken from https://math.stackexchange.com/a/2323106
-          // We do this to fake a progress that slows down as we process more
-          // events, as we don't know how many submodules there are upfront, or
-          // what does git have to do with them (cloning, just checking them
-          // out...)
-          percent: 1 - Math.exp(-submoduleEventCount * 0.25),
-        }
-      },
-    },
-    progress => {
-      const description =
-        progress.kind === 'progress' ? progress.details.text : progress.text
-
-      const value = progress.percent
-
-      progressCallback({
-        kind,
-        title,
-        description,
-        value,
-        remote: remote.name,
-      })
-    }
-  )
-
-  await git(args, repository.path, 'updateSubmodules', progressOpts)
-
-  // Final progress
-  progressCallback({
-    kind,
-    title,
-    description: 'Submodules updated',
-    value: 1,
-    remote: remote.name,
-  })
 }
 
 /**
@@ -208,12 +94,16 @@ export async function pull(
   await git(args, repository.path, 'pull', opts)
 
   // Update submodules after pull
-  await updateSubmodulesAfterPull(
+  const title = `Pulling ${remote.name}`
+  await updateSubmodulesAfterOperation(
     repository,
     remote,
     progressCallback
       ? clampProgress(PullStepWeight, 1, progressCallback)
       : undefined,
+    'pull',
+    title,
+    remote.name,
     allowFileProtocol
   )
 }
