+++ app/src/lib/git/submodule.ts
@@ -1,9 +1,128 @@
 import * as Path from 'path'
 
-import { git } from './core'
+import { git, IGitStringExecutionOptions } from './core'
 import { Repository } from '../../models/repository'
 import { SubmoduleEntry } from '../../models/submodule'
 import { pathExists } from '../../ui/lib/path-exists'
+import { executionOptionsWithProgress, IGitOutput } from '../progress'
+import {
+  envForRemoteOperation,
+  getFallbackUrlForProxyResolve,
+} from './environment'
+import { AuthenticationErrors } from './authentication'
+import { IRemote } from '../../models/remote'
+import { Progress } from '../../models/progress'
+
+/**
+ * Update submodules after a git operation.
+ *
+ * @param repository - The repository in which to update submodules
+ * @param remote - The remote for environment setup (can be null)
+ * @param progressCallback - An optional function which will be invoked
+ *                           with information about the current progress
+ *                           of the submodule update operation.
+ * @param progressKind - The kind of progress event ('checkout', 'pull', etc.)
+ * @param title - The title to use for progress reporting
+ * @param targetOrRemote - The target (for checkout) or remote name (for pull)
+ * @param allowFileProtocol - Whether to allow file:// protocol for submodules
+ */
+export async function updateSubmodulesAfterOperation<T extends Progress>(
+  repository: Repository,
+  remote: IRemote | null,
+  progressCallback: ((progress: T) => void) | undefined,
+  progressKind: T['kind'],
+  title: string,
+  targetOrRemote: string,
+  allowFileProtocol: boolean
+): Promise<void> {
+  const opts: IGitStringExecutionOptions = {
+    env: await envForRemoteOperation(
+      getFallbackUrlForProxyResolve(repository, remote)
+    ),
+    expectedErrors: AuthenticationErrors,
+  }
+
+  const args = [
+    ...(allowFileProtocol ? ['-c', 'protocol.file.allow=always'] : []),
+    'submodule',
+    'update',
+    '--init',
+    '--recursive',
+  ]
+
+  if (!progressCallback) {
+    await git(args, repository.path, 'updateSubmodules', opts)
+    return
+  }
+
+  // Initial progress
+  progressCallback({
+    kind: progressKind,
+    title,
+    description: 'Updating submodules',
+    value: 0,
+    // Add the target or remote field based on the progress kind
+    ...(progressKind === 'checkout'
+      ? { target: targetOrRemote }
+      : { remote: targetOrRemote }),
+  } as T)
+
+  let submoduleEventCount = 0
+
+  const progressOpts = await executionOptionsWithProgress(
+    { ...opts, trackLFSProgress: true },
+    {
+      parse(line: string): IGitOutput {
+        if (
+          line.match(/^Submodule path (.)+?: checked out /) ||
+          line.startsWith('Cloning into ')
+        ) {
+          submoduleEventCount += 1
+        }
+
+        return {
+          kind: 'context',
+          text: `Updating submodules: ${line}`,
+          // Math taken from https://math.stackexchange.com/a/2323106
+          // We do this to fake a progress that slows down as we process more
+          // events, as we don't know how many submodules there are upfront, or
+          // what does git have to do with them (cloning, just checking them
+          // out...)
+          percent: 1 - Math.exp(-submoduleEventCount * 0.25),
+        }
+      },
+    },
+    progress => {
+      const description =
+        progress.kind === 'progress' ? progress.details.text : progress.text
+
+      const value = progress.percent
+
+      progressCallback({
+        kind: progressKind,
+        title,
+        description,
+        value,
+        ...(progressKind === 'checkout'
+          ? { target: targetOrRemote }
+          : { remote: targetOrRemote }),
+      } as T)
+    }
+  )
+
+  await git(args, repository.path, 'updateSubmodules', progressOpts)
+
+  // Final progress
+  progressCallback({
+    kind: progressKind,
+    title,
+    description: 'Submodules updated',
+    value: 1,
+    ...(progressKind === 'checkout'
+      ? { target: targetOrRemote }
+      : { remote: targetOrRemote }),
+  } as T)
+}
 
 export async function listSubmodules(
   repository: Repository
